requirements.txt
anthropic>=0.25.0
pandas>=2.0.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
streamlit>=1.32.0
apache-airflow>=2.8.0
dbt-postgres>=1.7.0
scikit-learn>=1.4.0
numpy>=1.26.0
plotly>=5.20.0
python-dotenv>=1.0.0
